[FSDP] Start to generalize modules to ignore for mixed precision
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102010

The main use case here is that folks would like to ignore layer norm for mixed precision. This can now be enabled with:

```
mp_config = MixedPrecision(
            param_dtype=torch.float16,
            reduce_dtype=torch.float16,
            buffer_dtype=torch.float16,
            _mixed_precision_module_classes_to_ignore=[_BatchNorm, nn.LayerNorm],
        )
```

for LayerNorm, we also add module pre and post hooks to cast inputs to the appropriate full precision. This is not super generalizable at the moment, for example, if users want to skip module X, we won't install the same hooks for X, but we perhaps should. This can be done as follow up work for now as the main priority is to unblock layer norm skipping.
ghstack-source-id: 190148945

Differential Revision: [D46079957](https://our.internmc.facebook.com/intern/diff/D46079957/)